# api-docs.yaml
openapi: 3.0.1
info:
  title: EventService
  version: v0
servers:
- url: http://localhost:9013
  description: Event Service is a service that cares of maintaining events data
paths:
  /events/{eventId}:
    get:
      tags:
      - events-controller
      operationId: getEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
    put:
      tags:
      - events-controller
      operationId: createOrUpdateEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
      security:
      - bearer-key: []
    delete:
      tags:
      - events-controller
      operationId: deleteEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
      security:
      - bearer-key: []
    patch:
      tags:
      - events-controller
      operationId: updateEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
      security:
      - bearer-key: []
  /events:
    get:
      tags:
      - events-controller
      operationId: getAllEvents
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDTO'
    post:
      tags:
      - events-controller
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDTO'
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventResponseDTO'
      security:
      - bearer-key: []
components:
  schemas:
    ErrorResponseDTO:
      type: object
      properties:
        errorMessage:
          type: string
        timestamp:
          type: integer
          format: int64
    EventRequestDTO:
      type: object
      properties:
        name:
          type: string
        eventDate:
          type: string
        eventStartTime:
          type: string
        eventDurationInMinutes:
          type: string
        venue:
          type: string
        description:
          type: string
        artistNames:
          type: array
          items:
            type: string
    EventResponseDTO:
      type: object
      properties:
        eventId:
          type: integer
          format: int32
        name:
          type: string
        eventDate:
          type: string
        eventStartTime:
          type: string
        eventDurationInMinutes:
          type: integer
          format: int32
        venue:
          type: string
        description:
          type: string
        artistNames:
          type: array
          items:
            type: string
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT